requirements.txt
dj-database-url==0.5.0
Django==2.1.7
psycopg2-binary==2.8.5
pylint==2.5.2
python-engineio==3.13.0
python-http-client==3.2.7
requests==2.23.0
gunicorn
django-heroku
pylint-django
python-dotenv
selenium
nltk
tensorflow